requirements.txt
Flask==2.3.3
gunicorn==21.2.0
rembg==2.0.50
Pillow==10.0.1
python-multipart==0.0.6
asyncio==3.4.3
aiohttp==3.8.5
